new classes for parser and dumper
