+++ src/Viserio/Component/Parsers/Dumpers/IniDumper.php
@@ -4,34 +4,9 @@ namespace Viserio\Component\Parsers\Formats;
 
 use Viserio\Component\Contracts\Parsers\Dumper as DumperContract;
 use Viserio\Component\Contracts\Parsers\Exception\ParseException;
-use Viserio\Component\Contracts\Parsers\Format as FormatContract;
 
-class Ini implements FormatContract, DumperContract
+class IniDumper implements DumperContract
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function parse(string $payload): array
-    {
-        $ini = @parse_ini_string($payload, true, INI_SCANNER_RAW);
-
-        if (! $ini) {
-            $errors = error_get_last();
-
-            if ($errors === null) {
-                $errors['message'] = 'Invalid INI provided.';
-            }
-
-            throw new ParseException($errors);
-        }
-
-        foreach ($ini as $key => $value) {
-            $ini[$key] = self::normalize($value);
-        }
-
-        return $ini;
-    }
-
     /**
      * {@inheritdoc}
      */
