+++ src/Viserio/Component/Parsers/Dumpers/JsonDumper.php
@@ -1,0 +1,33 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Parsers\Formats;
+
+use Viserio\Component\Contracts\Parsers\Dumper as DumperContract;
+use Viserio\Component\Contracts\Parsers\Exception\DumpException;
+use Viserio\Component\Contracts\Parsers\Exception\ParseException;
+use Viserio\Component\Contracts\Parsers\Format as FormatContract;
+
+class JsonDumper implements DumperContract
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function dump(array $data): string
+    {
+        // Clear json_last_error()
+        json_encode(null);
+
+        $json = json_encode($data, JSON_PRETTY_PRINT);
+
+        if (json_last_error() !== JSON_ERROR_NONE) {
+            // @codeCoverageIgnoreStart
+            throw new DumpException(sprintf('JSON dumping failed: %s', json_last_error_msg()));
+            // @codeCoverageIgnoreEnd
+        }
+
+        $json = preg_replace('/\[\s+\]/', '[]', $json);
+        $json = preg_replace('/\{\s+\}/', '{}', $json);
+
+        return $json;
+    }
+}
