+++ src/Viserio/Component/Parsers/Dumpers/SerializeDumper.php
@@ -7,22 +7,8 @@ use Viserio\Component\Contracts\Parsers\Dumper as DumperContract;
 use Viserio\Component\Contracts\Parsers\Exception\ParseException;
 use Viserio\Component\Contracts\Parsers\Format as FormatContract;
 
-class Serialize implements FormatContract, DumperContract
+class SerializeDumper implements DumperContract
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function parse(string $payload): array
-    {
-        try {
-            return unserialize(trim($payload));
-        } catch (Throwable $exception) {
-            throw new ParseException([
-                'message' => 'Failed to parse serialized Data',
-            ]);
-        }
-    }
-
     /**
      * {@inheritdoc}
      */
