+++ src/Viserio/Component/Parsers/Dumpers/XmlDumper.php
@@ -1,0 +1,34 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Parsers\Formats;
+
+use DOMException;
+use InvalidArgumentException;
+use RuntimeException;
+use Spatie\ArrayToXml\ArrayToXml;
+use Viserio\Component\Contracts\Parsers\Dumper as DumperContract;
+use Viserio\Component\Contracts\Parsers\Exception\DumpException;
+use Viserio\Component\Contracts\Parsers\Exception\ParseException;
+use Viserio\Component\Contracts\Parsers\Format as FormatContract;
+use Viserio\Component\Parsers\Utils\XmlUtils;
+
+class XmlDumper implements DumperContract
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function dump(array $data): string
+    {
+        // @codeCoverageIgnoreStart
+        if (! class_exists(ArrayToXml::class)) {
+            throw new RuntimeException('Unable to dump XML, the ArrayToXml dumper is not installed.');
+        }
+        // @codeCoverageIgnoreEnd
+
+        try {
+            return ArrayToXml::convert($data);
+        } catch (DOMException $exception) {
+            throw new DumpException($exception->getMessage());
+        }
+    }
+}
